backend/requirements.txt
Django==2.1.5
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
rest-social-auth==2.0.2
django-cors-headers==2.4.0
django-rest-framework-social-oauth2==1.1.0
PyGithub==1.43.5